Added the beginnings of the wideCrate dDS client and correlator server.
This is copied from a dDSDemo script written by Taco, kenyoung, and simulates
the integration servers running on the SMA's correlator crate PowerPC's. This
will be modifed soon, however, to collect visibility data from the wideband
correlator's ROACH2s over 10 GbE and also for M&C of the new system.