Added the sma-walsh TBS plugin.
Currently this provides two functions. One for writing to a register
repeatedly using the PowerPC for testing driver-to-register latency
and jitter. And another for arming the SMA's SOWF/HB generator (which
requires a 'sync_ctrl' register to be present and more specifically
the 'sma_corr' bitcode to be programmed) using an absolute real time.