[wideCrate] Separated the Walsh sync procedure into its own function.
This allows it to now be called multiple times using pthread to
arm many ROACH's simultaneously. This is now done in the handle_package
function for two hard-coded ROACH2's.

Another change was to comment out the RPC server svc_freeargs call.
While trying to copy the package argument into a global variable so that
other threads could access it I kept getting seg. faults but couldn't
figure out why. The copy routine I wrote should have worked but the
package was being freed and there were still pointers pointing to that
freed memory somewhere! Instead of dealing with that I just removed the
freeing. Probably not the best approach but since this isn't actually
running as a server it's probably okay.