[dsm] Added the DSM TBS plugin for use at SMA.
This has some useful read/write functions for communicating with
DSM over katcp however the main purpose will for use within other
plugins and tcpborphserver itself.

This of course requires an installation of DSM on every machine
that's sharing variables, including the ROACH2's. For DSM the
documentation check out the SMA Wiki.