[scripts/catchdata] Updated the data catcher to process all direct-connect data.
The two direct-connect visibility IP addresses are currently hard-coded to the
values used at the SMA but these should work on the bench system too. The code now
properly waits for data from both ROACH2s (single sidband) for a single accumulation
before putting it on the queue.

Follow up commit will update the other end of the queue.