[scripts] Changed the default HB sync source from ADC0 to ADC1.
This is temporary and will eventually be changed back. Had to do this
because one of the direct-connect ROACH2s appears to have a broken cable
or contact leading from the front panel to ADC0 (normally reserved for
the HB). Paul had to swap HB and 1pps inputs which requires this commit's
changes to work.