[swarm] Began refactoring of software by adding a 'swarm' Python package.
This package presently contains 4 modules:

defines: contains global variables for all modules.

core: contains classes for controlling individual SWARM members and for
controlling SWARM as whole.

data: contains classes for listening for visibility data over 10 GbE, for
reordering the data, and for processing that data using user callbacks.
Also, the data handler presently sends data to the SMA dataCatcher and
corrSaver; this however may eventually be turned into a callback.

xeng: abstract classes for "baseline" and an X-engine class that uses a
Swarm instance as its mapping and outputs the output order in words of
the X-engine.

Also present is a script called "swarm_ctrl" that uses the new package and
can be used to setup and handle SWARM data. Note: this script will eventual-
ly take over the functionality of all the scripts in the scripts folder.