[swarm/core] Changed external sync pulse from HB to SOWF.
This implies a complete change to the way we do arm syncs. Firstly,
the sync can be done at any time so the DSM sync time is no longer
necessary as is the threaded calls to the sma_walsh sync arming
tbs_plugin. This commit just manually issues syncs by twiddling
the appropriate bits of sync_ctrl.

Along with this comes a hardware change to SWARM: the external HB
supplied has been changed to a pseudo-SOWF (multiple of a complete
Walsh cycle) generated by the DDS computer and transmitted over a
ribbon cable to a TTL buffer connected to the former HB distribution
box.