[tbs_plugins] Simplified the plugin Makefile.
It no longer tries to compile for both the current host and the
powerpc target, instead it just uses the CROSS_COMPILE option
which the user must set if they wish to cross compile.

To compile just the examples use:

$ make examples

This requires that KATCPDIR point to the location of katcp_devel.

To compile for SWARM use:

$ make swarm

Which requires the GLOBAL variable to be set (SMA only).

By default the all rule compiles only the examples.