[swarm/core] Added the ability to call SwarmMember methods from a Swarm instance.
Basically I am using Swarm.__getattr__ which is called when an attribute is not found
and checking to see if the attribute is callable and a method of SwarmMember. If so then
a callable is returned that passes *args and **kwargs to the appropriate method of each
of the valid SwarmMember's.

This means if you want to set the source to analog, instead of using the usual:

> swarm.members_do(lambda fid, m: m.set_source(2, 2))

You can instead just do:

> swarm.set_source(2, 2)

And it accomplishes the same thing. Introspection in IPython becomes of course a bit
confusing. Maybe I can resolve this later.