[swarm/dbe] The B-engine and DBE data catchers and handlers are now separate.
Since the two components data products are differently formatted I have renamed
DBEDataCatcher and DBEDataHandler to BengineDataCatcher and BengineDataHandler,
respectively. Then I added DBEDataCatcher and DBEDataHandler as sub-class of the
Bengine ones which extract the Bengine data from the VDIF frame output of the SWARM
DBE. Both data are then stored in the same numpy format.

Note: that the Bengine and DBE data outputs presently are different bitwidths, the
Bengine is 4-bits while the DBE is 2-bits. This difference is handle in the
respective handlers.