[callbacks/calibrate_vlbi] Added phase to the per-channel complex gain solver.
In other words it now uses the lag-space to find the delay *and* the phase of the solved
complex per-antenna, per-channel gain. Before it was just calculating the phase by
averaging the band and finding the angle of that. Now it actually finds the angle of the
peak in the lag-space.

This does not work very well since we will need to either pad the IFFT to get a finer
resolution on lag-space or as Laura's suggested just taking a sub-sample around the
peak lag and interpolating.