[callbacks/calibrate_vlbi] Used Laura's two-stage method for find the sub-sample delay and phase.
Basically we find the delay in lag-space as we normally would, then we slice out a
small section around the lag-space peak, the default case it's +/- 16 lags around the
peak. After that you take the FFT to get the frequency domain, then take a padded reverse
FFT to go back to an interpolated lag-space. The peak amplitude point now contains the
sub-sample delay in addition to the phase.