[swarm] Removed the chunk index in the SwarmDataPackage object.
This is redundant since the baseline index already includes the chunk information
and has been causing issues due to the fact that a chunk=0 only baseline will have
a chunk=1 data array. With the chunk index removed this is no longer an issue.

All callbacks that use this index have also been updated to reflect the change.