[swarm/dbe] Added a SwarmDBE class that will control the DBE (currently only configures interfaces).
This inherits from SwarmROACH just like SwarmMember does. Also had to rearrange some of the useful
interface imposter classes to all inherit from a common LocalInterface which itself inherits from an
abstract Interface class. Inheritance like this:

```
Interface -- LocalInterface --+--- DBEImposter
                              |
                               \__ MK6Imposter
```

The Imposter classes allow us to pretend to be a DBE (and thus receive B-engine packets) or pretend to
be a Mark6 (and thus receive the DBE packets if the DBE TX interfaces are on the switch) for debuggin
purposes.