[callbacks/calibrate_vlbi] Enabled the VLBI phase solver to solve BOTH chunks.
This involved moving most of the functionality of the solver from __call__ to
a new member, solve_for. This new member takes a "chunk" argument that specifies
which chunk to solve for in particular. Underneath it creates a correlation matrix
that only uses inputs for that chunk. The __call__ method now calls solve_for twice,
once for each chunk, and then merges the solutions for the servo loop to use.

Also in this commit the solver now reads a file defined by SWARM_PHASED_SUM_ANTENNAS
for which antennas in particular it should solve (and not solve) for.