[callbacks/calibrate_vlbi] Added single_chan option to CalibrateVLBI.
Basically when this is true (now the default) it will only solve for phase in a
single "pseudo-continuum" channel, which is the average of the entire band. When
this is false it will attempt solve for phase in every channel as well as delay
(really just rate-of-phase-change across the band).