[swarm/core] Added methods to SwarmROACH for handling tcpborphserver plugins.
This inclues a method for unloading, reloading, and listing all plugins. NOTE:
the SwarmROACH.reload_plugins command expects a Python ConfigParser file with
entries of the form:

   [plugin name]
   file: {required -- path to the plugin in the ROACH2 file system}
   init: {optional -- newline-delimited list of commands to run on init}

Any number of plugins can be listed and the order matters; they are loaded in
the order shown and unloaded in reverse order. This list should represent the
default set of plugins for SWARM.