[swarm/core] Swarm now returns a list of results when accessing SwarmMember methods.
In other words, when you want to call a particular SwarmMember method on all the valid
members in the Swarm object *AND* see the results of these you can still use the
swarm.{SwarmMember method} syntax but now a list of the individual results is returned.
For example, to get all the plugins currently loaded:

>>> swarm.plugin_list()
[[], [], [], [], [], [], [], []]

Which is the result for no plugins loaded.