[swarm] The bitcode used is now defined in the swarmMapping file.
The Swarm.load_mapping function now looks for lines which start with the
special parameter identifier, "!", and treats these as parameter mappings
that are then passed to the SwarmMember constructor. Currently the only one
that is really supported is the "bitcode" parameter which indicates which
bitcode to use on any particular SwarmMember.

NOTE: that once a parameter is seen once it's kept until it is overriden
by another "!" line. In other words one bitcode identifier at the start of the
swarmMapping file will set the bitcode for every FID defined in that file. This
also allows having different bitcodes per any arbitrary set of FIDs if needed.