[swarm] Migrated all scripts over to using Swarm instead of SwarmQuadrant.
Sadly this migration is not complete. Presently the swarm_ctrl script simply creates
a separate thread for each quadrant to run the data catchers in. This causes some issues,
one of which is it's no longer possible to kill the script with ctrl-C.

Upcoming commits will look to fix the above and have the catching code use Swarm
instead of SwarmQuadrant like everything else.