[swarm] Swarm.setup now takes a list of listener interfaces to assign to quadrants.
This can be a list of 1-Q interfaces where Q is the total number of quadrants. Each
quadrant will be assigned one interface from the list in order. If there is only one
interface in the list all quadrants will send data to that one interfaces. The default
at present is ['eth2', 'eth3'] which is intended to work well on the summit system.

NOTE: presently one interface mode is not possible due to how the DDR3 interleaving
works in the current bitcode. The next bitcode (for 10/11ths) will allow for interleaving
of visibility outputs across quadrants at which time we will switch to using only one
interface (for simplicity).