[swarm] SwarmDataCatcher now takes a real Swarm object instead of SwarmQuadrant.
This is the big commit that completes the migration to multiple quadrant SWARM
support. The SwarmDataCatcher class now handles the new Swarm object which is a container
for 1+ SwarmQuadrant objects. This means only *one* SwarmDataCatcher is needed so
the swarm_ctrl script needed to change appropriately (and no longer requires awkward
threading to get the job done).

This also slightly changes how the rawbacks are processed so those had to be updated.