[callbacks/sma_data] Convert data to lists before handing to send_integration.
This simple step reduces processing time by about half. The reason for this is that
PySequence_GetItem appears to be really slow at getting items from Numpy arrays,
converting them to lists first appears to speed this up greatly.

I suspect that sticking with ndarrays and using the Numpy C-API directly instead
would mean an even greater speed up but this remains to be seen.