[swarm/defines] Removed eth3 from default set of listener interfaces.
More than one interface is no longer needed due to the fact that different
quadrants now interleave their data using the visibs delay block. Thus, I am
removing the eth3 interface and leaving just eth2.

In the future this feature may be needed if the interleaving of four quadrants
takes too long and we want to compress the time it takes to send data from
SWARM. In that scenario we may want to use multiple receive interfaces and
interleave fewer quadrants at once.