[tbs_plugins/Makefile] Use the slalib library and include files in ACCAPP directory.
If you are not cross compiling you'll need to build slalib and place it in the
/application include and lib directories. If you're cross compiling this should
automatically use slalib that's installed on the ROACH2 application directory. If
building for a ROACH2 not at the SMA summit make sure to compile the slalib and place
it where the Makefile expects it to be.