[swarm/core] Established hierarchy of logger names for all swarm objects.
Each object now derives it's name using the getChild method of a parent logger.
At the top level is the module level logger which the Swarm object uses as its
parent, each SwarmQuadrant then uses the Swarm logger as a parent, and so forth.
This logging hierarchy makes more logical sense and allows for easier filtering
of log records.