[tbs_plugins/sma_dsm] Now the scan length AND progress are a single DSM structure.
This commit changes the DSM writing loop so that it writes to this structure instead of
the single floating point progress variable. NOTE: the structure contains two longs that
represent the current scan progress and scan length. These values are NOT floats and the
units of them are in 128*spectra_time where a spectra_time=2*channels/sample_rate.