[swarm/core] Added an idle function to the SwarmMember class.
This programs the FPGA on the member's ROACH2 with a so-called "idle" bitcode
which by default is called 'idle.bof'. The intent is for this bitcode to be a very
light-weight one which lowers the FPGA temperature and outputs nothing. This is only
needed at the moment since it appears that roach2.progdev('') doesn't actually un-
program the FPGA (based on looking at the temperatures and output 10 GbE rates).