[swarm] Handle conjugation properly, and fully, in SwarmXengine.
The X-engines naturally produce data that has certain baselines conjugated with respect
to the others (1-{678}, 2-{78}, and 3-8). The conjugation of these baselines was previously
done within SwarmDataPackage.set_data which was not only a terrible place to do it but also
introduced a bug whereby those baselines showed the LR and RL polarization cross-products to
be swapped.

This commit fixes the bug mentioned above and also simplifies the proper conjugation of
baselines by handling it in the X-engine reordering. A pleasant side effect is that this
speeds up processing by about 0.1 secs.