[swarm/data] Functionality of SwarmDataPackage.__init__ migrated to SwarmDataPackage.from_swarm.
The init function now takes simply a list of baselines instead of a Swarm instance. To create a data
package from a Swarm object use SwarmDataPackage.from_swarm instead. As before this is setup for adding
a new method that creates a data package from a string.

Note: this required a corresponding change in SwarmDataCatcher.