[swarm/data] Major change to SwarmDataCatcher: catch only ONE accumulation at a time.
This commit represents a major change in how the SwarmDataCatcher class handles receiving
data from all the X-engines. Major changes include:

1. Previously, the reorder thread attempted to handle multiple accumulations at once using
an accummulation buffer; now it only holds one at any time and will error out if data from
another accumulation is received before it's done processing the current accumulation. As a
result of this signaling between the catch and order threads is no longer needed and was
removed.

2. A couple new policies are now enforced with regards to the order and arrival time of the
incoming data. The data must start with the arrival of qid=0 and fid=0 and all data must arrive
within a predefined time (0.6 seconds); if this do not happen an error is propagated down to
the data handler.

3. The scan length (i.e. integration time) is now taken from the data themselves and not read
over DSM or katcp as before. A result of this is that all X-engines are checked to make sure
they provide the same scan length; if not, an error is propagated to the handler.

In practice, everything should work exactly the same as before but now the listening script
will crash with an error if the X-engines are not in sync or if some other problem messes with
the arrival of data and its order. When things are working properly there should be no difference
in behavior.