[swarm/core] Mapping files now have two types of configurable parameters; quadrant and member.
The member parameters, defined as before by a line with a '!' prefix, operate as they
normally would. A new type of parameter, defined by using a '$' prefix, will be translated
into a SwarmQuadrant attribute that will be set to whatever you define it to be in the mapping
file. For example, the line:

$ sdbe

Will result in that quadrant's SwarmQuadrant object having an attribute called 'sdbe' set to True.
If there is anything following the attribute name then it will be set to whatever string that is.