[swarm/core] Add Swarm-level get 2nd sideband beam phase method
At the Swarm level get_beamformer_second_sideband_phase will extract
phase for inputs grouped by quadrant, and then insert the returned
phases at the correct index corresponding to the inputs list passed
to the original call. This ensures that the returned phases are
matched to the requested inputs.

The returned phase for inputs not found in any quadrant default to
zero.