[ipython/magics] Add beamformer macros
Swarm shell now has the magics setbeam and showbeam available to allow
rapid setting and checking of beamformer configuration.

Set antennas in beam:
====================
%setbeam <antenna-list> [ gain=<gain-val> ]

<antenna-list> should evaluate to a list of antennas to include in the
beam. The same beam is formed in each polarization / sideband.

<gain-val> if supplied should evaluate to a scalar float. The same
gain is applied uniformly to the beam in each polarization / sideband.
If gain is not specified a value is automatically set, scaled by the
inverse of the length of the antenna list.

Examples:
// include antennas 1 through 8 in the beam
%setbeam range(8)
// include only antenna 2 in beam
%setbeam [2]
// include antennas 1,2,3 in beam and set gain to 5.125
%setbeam [1,2,3] gain=5.125

Show antennas in beam:
=====================
%showbeam

Prints a list of antennas in beam per chunk / sideband and shows the
gain applied to each sideband / polarization.